fix(auth): properly handle auth context
