+++ client/src/api/model/role.ts
@@ -8,7 +8,7 @@
 export type Role = typeof Role[keyof typeof Role];
 
 
-// eslint-disable-next-line @typescript-eslint/no-redeclare
+ 
 export const Role = {
   guest: 'guest',
   user: 'user',
