+++ client/src/auth/index.ts
@@ -1,1 +1,3 @@
-export { AuthProvider, useAuth, currentToken, setCurrentToken } from "./AuthContext";
+export { AuthProvider } from "./AuthProvider";
+export { useAuth } from "./useAuth";
+export { authStore } from "./authStore";
