+++ client/src/auth/useAuth.tsx
@@ -1,0 +1,11 @@
+import { useContext } from "react";
+import { AuthContext } from "./AuthContext";
+
+export function useAuth() {
+    const ctx = useContext(AuthContext);
+    if (!ctx) {
+        throw new Error("useAuth must be used within an AuthProvider");
+    }
+
+    return ctx;
+}
